+++ src/test/resources/results/testing/MyComponent9/testing_subpackage9_myComponent9_calcMotionCmds.h
@@ -16,7 +16,7 @@ double currentPositionX;
 double currentPositionY;
 double currentDirectionX;
 double currentDirectionY;
-double trimmedTrajectoryLength;
+int trimmedTrajectoryLength;
 Matrix trimmedTrajectoryX;
 Matrix trimmedTrajectoryY;
 double desiredDirectionX;
