+++ src/test/resources/results/testing/MyComponent9/testing_subpackage9_myComponent9_calcMotionCmds_driveToFirstPosition.h
@@ -11,7 +11,7 @@ public:
 bool isDriveToFirstPosition;
 double currentPositionX;
 double currentPositionY;
-double trimmedTrajectoryLength;
+int trimmedTrajectoryLength;
 Matrix trimmedTrajectoryX;
 Matrix trimmedTrajectoryY;
 double desiredDirectionX;
